Rename mypackage to mybot
