+++ src/mybot/__main__.py
@@ -1,0 +1,4 @@
+from . import main
+
+if __name__ == '__main__':
+    main()
