+++ src/mybot/cli.py
